backend/requirements.txt
groq
pydantic
requests
python-dotenv
fastapi
uvicorn
openai
twilio
